Refactor and reorganize code structure
- Removed deprecated files related to user and military functionalities.
- Updated documentation paths for official journals and user authentication.
- Implemented new controllers for user registration and login.
- Created use cases for querying official journals by municipality and state.
- Introduced new factories for creating controllers related to user and official journals.
- Enhanced error handling in controllers and use cases.
- Updated Swagger documentation for user authentication and official journals endpoints.
