+++ src/interfaces/docs/auth-user/auth-register-user.doc.ts
