+++ src/interfaces/docs/military/military-ott.doc.ts
