+++ src/interfaces/docs/official-journals/official-journals-state.doc.ts
