+++ src/interfaces/factories/auth-user/make-register-user-controller.ts
@@ -1,5 +1,5 @@
-import { AuthUserController } from '@interfaces/controllers/auth-user-controller';
-import { AuthUserUseCase } from '@usecases/user/auth-user-use-case';
+import { AuthUserController } from '@interfaces/controllers/auth-user/register-user-controller';
+import { AuthUserUseCase } from '@usecases/auth-user/register-user-use-case';
 import { PrismaAuthUserRepository } from '@infra/repositories/auth-user/auth-user-repository';
 
 export function makeAuthUserController(): AuthUserController {
