+++ src/interfaces/factories/military/make-military-ott-controller.ts
@@ -1,6 +1,6 @@
 import { ExercitoWebScraper } from "@infra/providers/gateways/exercito-work/exercito-web-scraper";
-import { ExercitoUseCase } from "@usecases/exercito-work/exercito-use-case";
-import { ExercitoController } from "@interfaces/controllers/exercito/exercito-controller";
+import { ExercitoUseCase } from "@usecases/military/military-ott-use-case";
+import { ExercitoController } from "@interfaces/controllers/military/military-ott-controller";
 
 export function makeExercitoController(): ExercitoController {
   const scraper = new ExercitoWebScraper();
