+++ src/interfaces/factories/official-journals/make-official-journals-state-controller.ts
@@ -1,6 +1,6 @@
 import { DiarioOficialEstadoWeb } from "@infra/providers/gateways/diario-oficial/diario-oficial-estado-web";
-import { ConsultarDiarioOficialEstadoUseCase } from "@usecases/diario-oficial/consultar-diario-oficial-estado";
-import { DiarioOficialEstadoController } from "@interfaces/controllers/diario-oficial/diario-oficial-estado-controller";
+import { ConsultarDiarioOficialEstadoUseCase } from "@usecases/official-journals/official-journals-state-use-case";
+import { DiarioOficialEstadoController } from "@interfaces/controllers/official-journals/official-journals-state-controller";
 
 export function makeDiarioOficialEstadoController(): DiarioOficialEstadoController {
     const diarioOficialEstadoWeb = new DiarioOficialEstadoWeb();
