+++ src/interfaces/routes/index.ts
@@ -1,13 +1,13 @@
 import { Router } from 'express';
 import { HomeController } from '../controllers/home-controller';
 import { enviarEmail } from '@interfaces/controllers/email/send-email-controller';
-import { makeUserController } from '@interfaces/factories/user/make-user-controller';
-import { makeExercitoController } from '@interfaces/factories/exercito/make-exercito-controller';
-import { makeDiarioOficialEstadoController } from '@interfaces/factories/diario-oficial/make-diario-oficial-estado-controller';
-import { makeDiarioOficialMunicipioController } from '@interfaces/factories/diario-oficial/make-diario-oficial-municipio-controller';
-import { makeAuthUserController } from '@interfaces/factories/auth-user/make-auth-user-controller';
+import { makeUserController } from '@interfaces/factories/user/make-create-user-controller';
+import { makeExercitoController } from '@interfaces/factories/military/make-military-ott-controller';
+import { makeDiarioOficialEstadoController } from '@interfaces/factories/official-journals/make-official-journals-state-controller';
+import { makeDiarioOficialMunicipioController } from '@interfaces/factories/official-journals/make-official-journals-municipality-controller';
+import { makeAuthUserController } from '@interfaces/factories/auth-user/make-register-user-controller';
 import { authenticateToken } from '@interfaces/middlewares/auth-middleware';
-import { makeLoginController } from '@interfaces/factories/auth-user/make-login-controller';
+import { makeLoginController } from '@interfaces/factories/auth-user/make-login-user-controller';
 
 
 export const router = Router();
